Debugger improvements
The log messages are more user friendly (not all event types are recognized).

Add items to a watchlist and monitor their states.
You can also send update the state from the watchlist by editing the text box and hitting Return:
this will send a command (or hit Esc to discard your changes).
Add an item to the watchlist by clicking its name in the log.

Signed-off-by: Yannick Schaus <habpanel@schaus.net>